add useFavoriteHandler to offload favorite handling from cryptodata component
